converted some video and fwipe fields to int16

+++ am_map.c
@@ -287,10 +287,10 @@ extern boolean viewactive;
 
 void
 V_MarkRect
-( int32_t	x,
-  int32_t	y,
-  int32_t	width,
-  int32_t	height );
+( int16_t	x,
+  int16_t	y,
+  int16_t	width,
+  int16_t	height );
